Add support for rendering with a URL prefix
This adds an optional MEDUSA_URL_PREFIX setting option that causes Django's URL
reversing to render URLS prefixed with this string. This is necessary when
hosting Django projects on a URI path other than /, as a proper WSGI environment
is not present to tell Django what URL prefix to use.